# IADocumentClassifier.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IA Document Classifier API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para consultar todos los clientes",
        "operationId": "GetAll",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "identification",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para Crear clientes",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para Actualizar clientes",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para consultar por cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Metodo para eliminar clientes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para consultar todos los Roles",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para crear Roles",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para actulizar Roles",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para consultar por Roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para Eliminar o borrar Roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {
          "type": "object",
          "additionalProperties": false
        }
      },
      "ClientsDTO": {
        "type": "object",
        "properties": {
          "client_Id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "identification": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "boolean"
          },
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RolesDTO": {
        "type": "object",
        "properties": {
          "rol_Id": {
            "type": "integer",
            "format": "int32"
          },
          "rolName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}